# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Keep this title stable: the generated package paths depend on it.
  title: Api
  version: 1.0.0
  description: Rakt Kavach National Blood Grid API
servers:
  - url: /api
    description: Shared API service
...
paths:
...
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboards]
      summary: Get a role-aware command center summary
      parameters:
        - $ref: "#/components/parameters/Role"
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /donors:
    get:
      operationId: listDonors
      tags: [donors]
      summary: Search the donor registry
      parameters:
        - $ref: "#/components/parameters/BloodGroup"
        - name: city
          in: query
          schema: { type: string }
        - name: eligible
          in: query
          schema: { type: boolean }
        - name: search
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Donor registry results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonorListResponse"
    post:
      operationId: createDonor
      tags: [donors]
      summary: Register a donor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonorInput"
      responses:
        "201":
          description: Donor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
  /donors/{donorId}:
    get:
      operationId: getDonor
      tags: [donors]
      summary: Get a donor profile
      parameters:
        - $ref: "#/components/parameters/DonorId"
      responses:
        "200":
          description: Donor profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      operationId: updateDonor
      tags: [donors]
      summary: Update a donor profile
      parameters:
        - $ref: "#/components/parameters/DonorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DonorUpdate"
      responses:
        "200":
          description: Updated donor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Donor"
        "404": { $ref: "#/components/responses/NotFound" }
  /donors/{donorId}/history:
    get:
      operationId: getDonorHistory
      tags: [donors]
      summary: Get a donor's donation history
      parameters:
        - $ref: "#/components/parameters/DonorId"
      responses:
        "200":
          description: Donation history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DonationHistoryResponse"
  /inventory:
    get:
      operationId: listBloodUnits
      tags: [inventory]
      summary: List live blood inventory
      parameters:
        - $ref: "#/components/parameters/BloodGroup"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/InventoryStatus"
        - name: facilityId
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 100 }
      responses:
        "200":
          description: Blood units
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodUnitListResponse"
    post:
      operationId: createBloodUnit
      tags: [inventory]
      summary: Add a verified blood unit to inventory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodUnitInput"
      responses:
        "201":
          description: Blood unit created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodUnit"
  /inventory/{unitId}:
    patch:
      operationId: updateBloodUnit
      tags: [inventory]
      summary: Update a blood unit
      parameters:
        - $ref: "#/components/parameters/UnitId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodUnitUpdate"
      responses:
        "200":
          description: Updated blood unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodUnit"
        "404": { $ref: "#/components/responses/NotFound" }
  /inventory/timeline:
    get:
      operationId: getBloodUnitTimeline
      tags: [inventory]
      summary: Get recent inventory events
      parameters:
        - name: facilityId
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Inventory timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodUnitTimelineResponse"
  /blood-requests:
    get:
      operationId: listBloodRequests
      tags: [requests]
      summary: List active blood requests
      parameters:
        - $ref: "#/components/parameters/BloodGroup"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/RequestStatus"
        - name: facilityId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Blood requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequestListResponse"
    post:
      operationId: createBloodRequest
      tags: [requests]
      summary: Create a blood request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodRequestInput"
      responses:
        "201":
          description: Blood request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"
  /blood-requests/{requestId}/status:
    patch:
      operationId: updateBloodRequestStatus
      tags: [requests]
      summary: Update blood request status
      parameters:
        - $ref: "#/components/parameters/RequestId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BloodRequestUpdateStatus"
      responses:
        "200":
          description: Updated request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BloodRequest"
        "404": { $ref: "#/components/responses/NotFound" }
  /emergencies:
    get:
      operationId: listEmergencies
      tags: [emergencies]
      summary: List SOS emergency cases
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/EmergencyStatus"
      responses:
        "200":
          description: Emergency cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyListResponse"
    post:
      operationId: createEmergency
      tags: [emergencies]
      summary: Create an SOS emergency case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyInput"
      responses:
        "201":
          description: Emergency created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Emergency"
  /emergencies/{emergencyId}/status:
    patch:
      operationId: updateEmergencyStatus
      tags: [emergencies]
      summary: Update an emergency status
      parameters:
        - $ref: "#/components/parameters/EmergencyId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmergencyUpdateStatus"
      responses:
        "200":
          description: Updated emergency
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Emergency"
  /emergencies/trends:
    get:
      operationId: getEmergencyTrends
      tags: [emergencies]
      summary: Get emergency trend data
      parameters:
        - name: period
          in: query
          schema:
            $ref: "#/components/schemas/TrendPeriod"
      responses:
        "200":
          description: Emergency trends
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmergencyTrendResponse"
  /facilities:
    get:
      operationId: listFacilities
      tags: [facilities]
      summary: List blood network facilities
      parameters:
        - name: type
          in: query
          schema:
            $ref: "#/components/schemas/FacilityType"
        - name: city
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Facilities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FacilityListResponse"
  /verification/qr/{token}:
    get:
      operationId: verifyQrPass
      tags: [verification]
      summary: Verify a donor or blood unit QR pass
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string, minLength: 8 }
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrVerification"
        "404": { $ref: "#/components/responses/NotFound" }
...
  /audit-logs:
    get:
      operationId: getAuditLogs
      tags: [audit]
      summary: Get immutable operational audit events
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        "200":
          description: Audit events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogListResponse"
...
components:
  parameters:
    DonorId:
      name: donorId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
    UnitId:
      name: unitId
      in: path
      required: true
      schema: { type: string, minLength: 1 }
...
    BloodGroup:
      name: bloodGroup
      in: query
      schema:
        $ref: "#/components/schemas/BloodGroup"
    Role:
      name: role
      in: query
      schema:
        $ref: "#/components/schemas/UserRole"
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    InventoryStatus:
      type: string
      enum: [available, reserved, quarantined, transfused, expired]
    RequestStatus:
      type: string
      enum: [open, matched, fulfilled, cancelled]
...
    TrendPeriod:
      type: string
      enum: [24h, 7d, 30d, 90d]
    Donor:
      type: object
      required: [id, donorCode, bloodGroup, city, eligibilityStatus, createdAt]
      properties:
        id: { type: string }
        donorCode: { type: string }
        fullName: { type: string }
        bloodGroup: { $ref: "#/components/schemas/BloodGroup" }
        city: { type: string }
        phoneLast4: { type: ["string", "null"] }
        eligibilityStatus:
          type: string
          enum: [eligible, due-soon, ineligible, unknown]
        lastDonationAt: { type: ["string", "null"], format: date-time }
        nextEligibleAt: { type: ["string", "null"], format: date-time }
        rewardTier: { type: ["string", "null"] }
        bloodCredits: { type: integer, minimum: 0 }
        verified: { type: boolean }
        createdAt: { type: string, format: date-time }
    DonorInput:
      type: object
      required: [fullName, bloodGroup, city]
      properties:
        fullName: { type: string, minLength: 2 }
        bloodGroup: { $ref: "#/components/schemas/BloodGroup" }
        city: { type: string, minLength: 2 }
        phoneLast4: { type: string, minLength: 4, maxLength: 4 }
    DonorUpdate:
      type: object
      properties:
        fullName: { type: string, minLength: 2 }
        city: { type: string, minLength: 2 }
        phoneLast4: { type: string, minLength: 4, maxLength: 4 }
        eligibilityStatus:
          type: string
          enum: [eligible, due-soon, ineligible, unknown]
        nextEligibleAt: { type: ["string", "null"], format: date-time }
        rewardTier: { type: ["string", "null"] }
        bloodCredits: { type: integer, minimum: 0 }
    DonorListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Donor" } }
        total: { type: integer }
...
    DonationHistoryResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/DonationHistory" } }
        total: { type: integer }
...
    BloodUnitUpdate:
      type: object
      properties:
        status: { $ref: "#/components/schemas/InventoryStatus" }
        facilityId: { type: string }
        qrToken: { type: ["string", "null"] }
    BloodUnitListResponse:
      type: object
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/BloodUnit" } }
        total: { type: integer }
...
    DashboardSummary:
      type: object
      required: [role, metrics, bloodGroupCounts, recentActivity]
      properties:
        role: { $ref: "#/components/schemas/UserRole" }
        metrics:
          type: object
          additionalProperties: { type: number }
        bloodGroupCounts:
          type: array
          items:
            type: object
            required: [bloodGroup, availableUnits, openRequests]
            properties:
              bloodGroup: { $ref: "#/components/schemas/BloodGroup" }
              availableUnits: { type: integer }
              openRequests: { type: integer }
        recentActivity:
          type: array
          items: { $ref: "#/components/schemas/AuditLog" }